Refined styling and added placeholder module data
