+++ frontend/src/hooks/useUserModules.ts
@@ -1,22 +1,5 @@
 import { useEffect, useState } from 'react';
-
-export interface UserModule {
-    moduleId: string;
-    type: ModuleType;
-    // TODO: do I want to also add a completed field to the schema?
-    isOpen: boolean;
-}
-
-// TODO: need to add this to the types file
-export enum ModuleType {
-    // TODO: make sure this enum is adjusted to the right vocab in backend
-    CONCEPT_GUIDE = "concept_guide",
-    PYTHON_GUIDE = "python_guide",
-    RECOGNITION_GUIDE = "recognition_guide",
-    QUIZ = "quiz",
-    CHALLENGE = "challenge",
-    CHALLENGE_SOLUTION = "solution_guide",
-}   
+import { UserModule } from '../types/ModuleTypes';
 
 const useUserModules = (unitId: number) => {
     const [userModules, setUserModules] = useState<UserModule[]>([]);
