+++ frontend/src/styles/Unit.module.css
@@ -2,6 +2,7 @@
     padding: 50px;
     display: flex;
     flex-direction: row;
+    justify-content: center;
     gap: 80px;
 }
